Made expires type int64 to support long ban times

+++ internal/api/types/VerificationStatus.go
@@ -9,7 +9,7 @@ type VerificationStatus struct {
 	Account_id      string                       `json:"account_id,omitempty"`
 	Attributes      []string                     `json:"attributes,omitempty"`
 	Ban_reason      string                       `json:"ban_reason,omitempty"`
-	Expires         int                          `json:"expires,omitempty"`
+	Expires         int64                        `json:"expires,omitempty"`
 	Is_primary      bool                         `json:"is_primary,omitempty"`
 	Primary_user_id string                       `json:"primary_user_id,omitempty"`
 	Service_links   []ServiceLink                `json:"service_links,omitempty"`
